requirements.txt
genologics
click
pyyaml
requests
